+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/SampleSlide.java
@@ -1,5 +1,7 @@
 package uk.ac.bbsrc.tgac.miso.core.data;
 
+import java.math.BigDecimal;
+
 public interface SampleSlide extends SampleTissueProcessing {
 
   public static final String SUBCATEGORY_NAME = "Slide";
@@ -23,5 +25,21 @@ public interface SampleSlide extends SampleTissueProcessing {
   public Stain getStain();
 
   public void setStain(Stain stain);
+
+  public BigDecimal getPercentTumour();
+
+  public void setPercentTumour(BigDecimal percentTumour);
+
+  public BigDecimal getPercentNecrosis();
+
+  public void setPercentNecrosis(BigDecimal percentNecrosis);
+
+  public BigDecimal getMarkedArea();
+
+  public void setMarkedArea(BigDecimal markedArea);
+
+  public BigDecimal getMarkedAreaPercentTumour();
+
+  public void setMarkedAreaPercentTumour(BigDecimal markedAreaPercentTumour);
 
 }